Fixing the build
https://jira.ccpgames.com/browse/GFX-1456

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/eve/src/pyFSD/": change = 1069791]

+++ FsdDictOffsetLookup.cpp
@@ -127,7 +127,7 @@ SemiFastStringLoadingHashedOffsetLookup::SemiFastStringLoadingHashedOffsetLookup
 		fos.offset = offset;
 		fos.size = size;
 		m_hash[key] = fos;
-		offsetToStartOfData += offsetToKey + 4 + key.length();
+		offsetToStartOfData += offsetToKey + 4 + uint32_t(key.length());
 	}	
 }
